requirements.txt
# Core dependencies (all available in the `envr` conda environment)
pyarrow>=14.0
rasterio>=1.3
torch>=2.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.65
